Fixed circular dependency problem by merging dataprep and manhattanization files
